+++ TencentCloud/Lcic/V20220817/Models/DescribeRoomResponse.cs
@@ -181,7 +181,7 @@ namespace TencentCloud.Lcic.V20220817.Models
         public long? EndDelayTime{ get; set; }
 
         /// <summary>
-        /// 直播方式：0 常规模式（默认）1 回放直播模式（伪直播）
+        /// 直播类型：0 常规（默认）1 伪直播
         /// </summary>
         [JsonProperty("LiveType")]
         public ulong? LiveType{ get; set; }