+++ TencentCloud/Ocr/V20181119/Models/ElectronicAirTransportDetail.cs
@@ -1,0 +1,108 @@
+/*
+ * Copyright (c) 2018 THL A29 Limited, a Tencent company. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+namespace TencentCloud.Ocr.V20181119.Models
+{
+    using Newtonsoft.Json;
+    using System.Collections.Generic;
+    using TencentCloud.Common;
+
+    public class ElectronicAirTransportDetail : AbstractModel
+    {
+        
+        /// <summary>
+        /// 航段序号
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("FlightSegment")]
+        public string FlightSegment{ get; set; }
+
+        /// <summary>
+        /// 始发站
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("StationGetOn")]
+        public string StationGetOn{ get; set; }
+
+        /// <summary>
+        /// 目的站
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("StationGetOff")]
+        public string StationGetOff{ get; set; }
+
+        /// <summary>
+        /// 承运人
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("Carrier")]
+        public string Carrier{ get; set; }
+
+        /// <summary>
+        /// 航班号
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("FlightNumber")]
+        public string FlightNumber{ get; set; }
+
+        /// <summary>
+        /// 座位等级
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("SeatLevel")]
+        public string SeatLevel{ get; set; }
+
+        /// <summary>
+        /// 承运日期
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("FlightDate")]
+        public string FlightDate{ get; set; }
+
+        /// <summary>
+        /// 起飞时间
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("DepartureTime")]
+        public string DepartureTime{ get; set; }
+
+        /// <summary>
+        /// 客票级别/客票类别
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("FareBasis")]
+        public string FareBasis{ get; set; }
+
+
+        /// <summary>
+        /// For internal usage only. DO NOT USE IT.
+        /// </summary>
+        public override void ToMap(Dictionary<string, string> map, string prefix)
+        {
+            this.SetParamSimple(map, prefix + "FlightSegment", this.FlightSegment);
+            this.SetParamSimple(map, prefix + "StationGetOn", this.StationGetOn);
+            this.SetParamSimple(map, prefix + "StationGetOff", this.StationGetOff);
+            this.SetParamSimple(map, prefix + "Carrier", this.Carrier);
+            this.SetParamSimple(map, prefix + "FlightNumber", this.FlightNumber);
+            this.SetParamSimple(map, prefix + "SeatLevel", this.SeatLevel);
+            this.SetParamSimple(map, prefix + "FlightDate", this.FlightDate);
+            this.SetParamSimple(map, prefix + "DepartureTime", this.DepartureTime);
+            this.SetParamSimple(map, prefix + "FareBasis", this.FareBasis);
+        }
+    }
+}
+