+++ TencentCloud/Ocr/V20181119/Models/ElectronicTrainTicket.cs
@@ -1,0 +1,228 @@
+/*
+ * Copyright (c) 2018 THL A29 Limited, a Tencent company. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+namespace TencentCloud.Ocr.V20181119.Models
+{
+    using Newtonsoft.Json;
+    using System.Collections.Generic;
+    using TencentCloud.Common;
+
+    public class ElectronicTrainTicket : AbstractModel
+    {
+        
+        /// <summary>
+        /// 购方名称
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("BuyerName")]
+        public string BuyerName{ get; set; }
+
+        /// <summary>
+        /// 购方识别号
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("BuyerTaxCode")]
+        public string BuyerTaxCode{ get; set; }
+
+        /// <summary>
+        /// 发票号码
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("Number")]
+        public string Number{ get; set; }
+
+        /// <summary>
+        /// 开票日期
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("Date")]
+        public string Date{ get; set; }
+
+        /// <summary>
+        /// 价税合计（中文大写）
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("TotalCN")]
+        public string TotalCN{ get; set; }
+
+        /// <summary>
+        /// 税额
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("Tax")]
+        public string Tax{ get; set; }
+
+        /// <summary>
+        /// 业务类型，0：退票，1:售票
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ServiceType")]
+        public string ServiceType{ get; set; }
+
+        /// <summary>
+        /// 出发时间
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("TimeGetOn")]
+        public string TimeGetOn{ get; set; }
+
+        /// <summary>
+        /// 车次
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("TrainNumber")]
+        public string TrainNumber{ get; set; }
+
+        /// <summary>
+        /// 发票代码
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("Code")]
+        public string Code{ get; set; }
+
+        /// <summary>
+        /// 席别
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("SeatType")]
+        public string SeatType{ get; set; }
+
+        /// <summary>
+        /// 乘车日期
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("DateGetOn")]
+        public string DateGetOn{ get; set; }
+
+        /// <summary>
+        /// 车厢
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("TrainCabin")]
+        public string TrainCabin{ get; set; }
+
+        /// <summary>
+        /// 出发站
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("StationGetOn")]
+        public string StationGetOn{ get; set; }
+
+        /// <summary>
+        /// 电子客票号
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("ElectronicNumber")]
+        public string ElectronicNumber{ get; set; }
+
+        /// <summary>
+        /// 姓名
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("PassengerName")]
+        public string PassengerName{ get; set; }
+
+        /// <summary>
+        /// 证件号
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("PassengerNo")]
+        public string PassengerNo{ get; set; }
+
+        /// <summary>
+        /// 金额
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("Amount")]
+        public string Amount{ get; set; }
+
+        /// <summary>
+        /// 到达站
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("StationGetOff")]
+        public string StationGetOff{ get; set; }
+
+        /// <summary>
+        /// 税率
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("TaxRate")]
+        public string TaxRate{ get; set; }
+
+        /// <summary>
+        /// 席位
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("Seat")]
+        public string Seat{ get; set; }
+
+        /// <summary>
+        /// 价税合计
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("Total")]
+        public string Total{ get; set; }
+
+        /// <summary>
+        /// 校验码
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("CheckCode")]
+        public string CheckCode{ get; set; }
+
+        /// <summary>
+        /// 发票状态代码，0正常 1 未更新  2作废 3已红冲
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("StateCode")]
+        public string StateCode{ get; set; }
+
+
+        /// <summary>
+        /// For internal usage only. DO NOT USE IT.
+        /// </summary>
+        public override void ToMap(Dictionary<string, string> map, string prefix)
+        {
+            this.SetParamSimple(map, prefix + "BuyerName", this.BuyerName);
+            this.SetParamSimple(map, prefix + "BuyerTaxCode", this.BuyerTaxCode);
+            this.SetParamSimple(map, prefix + "Number", this.Number);
+            this.SetParamSimple(map, prefix + "Date", this.Date);
+            this.SetParamSimple(map, prefix + "TotalCN", this.TotalCN);
+            this.SetParamSimple(map, prefix + "Tax", this.Tax);
+            this.SetParamSimple(map, prefix + "ServiceType", this.ServiceType);
+            this.SetParamSimple(map, prefix + "TimeGetOn", this.TimeGetOn);
+            this.SetParamSimple(map, prefix + "TrainNumber", this.TrainNumber);
+            this.SetParamSimple(map, prefix + "Code", this.Code);
+            this.SetParamSimple(map, prefix + "SeatType", this.SeatType);
+            this.SetParamSimple(map, prefix + "DateGetOn", this.DateGetOn);
+            this.SetParamSimple(map, prefix + "TrainCabin", this.TrainCabin);
+            this.SetParamSimple(map, prefix + "StationGetOn", this.StationGetOn);
+            this.SetParamSimple(map, prefix + "ElectronicNumber", this.ElectronicNumber);
+            this.SetParamSimple(map, prefix + "PassengerName", this.PassengerName);
+            this.SetParamSimple(map, prefix + "PassengerNo", this.PassengerNo);
+            this.SetParamSimple(map, prefix + "Amount", this.Amount);
+            this.SetParamSimple(map, prefix + "StationGetOff", this.StationGetOff);
+            this.SetParamSimple(map, prefix + "TaxRate", this.TaxRate);
+            this.SetParamSimple(map, prefix + "Seat", this.Seat);
+            this.SetParamSimple(map, prefix + "Total", this.Total);
+            this.SetParamSimple(map, prefix + "CheckCode", this.CheckCode);
+            this.SetParamSimple(map, prefix + "StateCode", this.StateCode);
+        }
+    }
+}
+