+++ TencentCloud/Tione/V20211111/Models/DescribeBillingSpecsRequest.cs
@@ -25,16 +25,16 @@ namespace TencentCloud.Tione.V20211111.Models
     {
         
         /// <summary>
-        /// 枚举值：TRAIN、NOTEBOOK、INFERENCE
-        /// </summary>
-        [JsonProperty("TaskType")]
-        public string TaskType{ get; set; }
-
-        /// <summary>
         /// 付费模式：POSTPAID_BY_HOUR按量计费、PREPAID包年包月
         /// </summary>
         [JsonProperty("ChargeType")]
         public string ChargeType{ get; set; }
+
+        /// <summary>
+        /// 枚举值：TRAIN、NOTEBOOK、INFERENCE
+        /// </summary>
+        [JsonProperty("TaskType")]
+        public string TaskType{ get; set; }
 
         /// <summary>
         /// 资源类型：CALC 计算资源、CPU CPU资源、GPU GPU资源、CBS云硬盘
@@ -48,8 +48,8 @@ namespace TencentCloud.Tione.V20211111.Models
         /// </summary>
         public override void ToMap(Dictionary<string, string> map, string prefix)
         {
-            this.SetParamSimple(map, prefix + "TaskType", this.TaskType);
             this.SetParamSimple(map, prefix + "ChargeType", this.ChargeType);
+            this.SetParamSimple(map, prefix + "TaskType", this.TaskType);
             this.SetParamSimple(map, prefix + "ResourceType", this.ResourceType);
         }
     }